refactor(json): rename string json readers
