+++ java/fory-json/src/main/java/org/apache/fory/json/codec/BaseObjectCodec.java
@@ -34,8 +34,8 @@ import org.apache.fory.json.meta.JsonFieldAccessor;
 import org.apache.fory.json.meta.JsonFieldInfo;
 import org.apache.fory.json.meta.JsonFieldTable;
 import org.apache.fory.json.reader.JsonReader;
-import org.apache.fory.json.reader.Latin1StringJsonReader;
-import org.apache.fory.json.reader.Utf16StringJsonReader;
+import org.apache.fory.json.reader.Latin1JsonReader;
+import org.apache.fory.json.reader.Utf16JsonReader;
 import org.apache.fory.json.reader.Utf8JsonReader;
 import org.apache.fory.json.resolver.JsonTypeInfo;
 import org.apache.fory.json.resolver.JsonTypeResolver;
@@ -229,13 +229,13 @@ public abstract class BaseObjectCodec extends AbstractJsonCodec {
 
   @Internal
   public Object readLatin1NonNull(
-      Latin1StringJsonReader reader, JsonTypeInfo typeInfo, JsonTypeResolver resolver) {
+      Latin1JsonReader reader, JsonTypeInfo typeInfo, JsonTypeResolver resolver) {
     return readNonNull(reader, typeInfo, resolver);
   }
 
   @Internal
   public Object readUtf16NonNull(
-      Utf16StringJsonReader reader, JsonTypeInfo typeInfo, JsonTypeResolver resolver) {
+      Utf16JsonReader reader, JsonTypeInfo typeInfo, JsonTypeResolver resolver) {
     return readNonNull(reader, typeInfo, resolver);
   }
